BoxView, Layout, Logo, NavBar re-written in styled-components. Removed Footer folder, won't be required for now. Removed the styles from Inputs in CustomizeBackground/Text and will style those seperately
