+++ components/CustomizeText/index.js
@@ -1,11 +1,9 @@
-import React, { useEffect, useState } from "react";
-
-import styles from "./index.module.css";
+import React from "react";
 
 function CustomizeText({ box, setBox }) {
   return (
     <>
-      <label htmlFor='set-background-color' className={styles.input}>
+      <label htmlFor='set-background-color'>
         Edit Text
         <input
           type='text'
