--- components/CustomizeText/index.module.css
@@ -1,6 +1,0 @@
-.input {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
-}
