--- components/Footer/index.module.css
@@ -1,14 +1,0 @@
-.footer {
-    background-color: white;
-    z-index: 1000000;
-    font-size: 12px;
-    width: inherit;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.footerLink {
-    margin: 2px;
-    padding: 0px 5px;
-}
