+++ components/Layout/Layout.js
@@ -1,0 +1,10 @@
+import NavBar from "../NavBar";
+
+export default function Layout({ className, children }) {
+  return (
+    <>
+      <NavBar />
+      <main className={className}>{children}</main>
+    </>
+  );
+}
