+++ components/Layout/Layout.style.js
@@ -1,0 +1,10 @@
+import styled from "styled-components"
+import Layout from "./Layout"
+
+export const StyledLayout = styled(Layout)`
+    display: grid;
+    justify-items: center;
+    align-items: center;
+    width: inherit;
+    height: 100%;
+`
