--- components/Layout/index.module.css
@@ -1,7 +1,0 @@
-.main {
-    display: grid;
-    justify-items: center;
-    align-items: center;
-    width: inherit;
-    height: 100%;
-}
