--- components/Logo/index.module.css
@@ -1,18 +1,0 @@
-.logoContainer {
-    height: 20px;
-    font-size: 18px;
-    font-weight: 500;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.logoImage {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.logoImage img {
-    width: 35px;
-}
