+++ components/NavBar/index.js
@@ -1,28 +1,3 @@
-import React, { useEffect, useState } from "react";
-
-import styles from "./index.module.css";
-
-import Logo from "../Logo";
-import SignInButton from "../SignInButton";
-
-import { useUserStore } from "../../stores/userStore";
-import SignOutButton from "../SignOutButton";
-
-function NavBar() {
-  const [isUser, setIsUser] = useState({});
-
-  const user = useUserStore((state) => state.user);
-
-  useEffect(() => {
-    setIsUser(user);
-  }, [user]);
-
-  return (
-    <header className={styles.navBar}>
-      <Logo />
-      {isUser ? <SignOutButton /> : <SignInButton />}
-    </header>
-  );
-}
+import { StyledNavBar as NavBar } from "./NavBar.style"
 
 export default NavBar;
